pyproject.toml
[tool.poetry]
name = "user-registration-api"
version = "0.1.0"
description = ""
authors = ["el-collins <charlescollins476@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
passlib = "^1.7.4"
uvicorn = "^0.29.0"
motor = "^3.4.0"
httpx = "^0.27.0"
bcrypt = "4.0.1"
pydantic = {extras = ["email"], version = "^2.6.4"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
